[FIX] account: prioritize env.company on creation of account.move
The current behavior is to take the company of the journal when creating
a move. This makes sense when working in multi-company but not in
multi-branch.
For instance:
* Have two companies, one is the branch of the other.
* Have a Sale journal in the main company, and no journal in the branch
* While having both companies activated, but logged in the branch,
  create a Customer Invoice in the Sale journal

This was setting the invoice in the main company instead of in the
branch, which is what users are used to as a default value everywhere
else (because it is the main activated company).

closes odoo/odoo#185511

Related: odoo/enterprise#72989
Signed-off-by: Raphael Collet <rco@odoo.com>