[IMP] marketing_card: allow picking a render language
Currently cards are rendered in the language of the current user.
Additionally translations are only performed for fields and not for static content.

We now allow each record to have 1 card per language for any given card campaign.
Users are now able to pick a language for their card mailing to be used
when rendering the cards. If the mailing is for a card campaign
it MUST have a language set to select which set of cards should
be linked to the recipients.

This allows card recipients to better promote their talks locally.

Statistics will count 1 card per recipient even when they have multiple
cards in different languages. Only the card with the highest status is
counted towards statistics ("exists" < "opened" < "shared").

task-4247003

Part-of: odoo/odoo#201197
Related: odoo/upgrade#7365
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>