[ADD] marketing_card_event: bridge for event modules
We want to have some more custom behavior for cards on models linked to events.
To avoid adding many bridge modules this is implemented in a mostly generic way
as the behavior isn't so specific.

- Events now have a "Send Cards" button opening the campaign mailing wizard
- That wizard lets users pick a campaign (rather than being readonly)
- Updating the campaign updates the preview card in the body of the mailing
- Picking an event-related model sets the share url to the share url
of the event of the preview record when possible

task-4247003

Part-of: odoo/odoo#201197
Related: odoo/upgrade#7365
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>