[IMP] orm: Model._access_domain() replaces Model._check_access()
Replace method _check_access() with _access_domain(), which returns the
domain that defines the accessible records.

When the user has no model access, the method returns the FALSE domain.
Otherwise, it returns the record rules for the model.

Method _make_access_error_message() is provided to generate the error
message.

Part-of: odoo/odoo#254381
Related: odoo/enterprise#110965
Signed-off-by: Raphael Collet <rco@odoo.com>