[IMP] google_calendar: addition operator for GoogleEvent
Before this commit, it wasn't possible to add two instances
of GoogleEvent to combine them into a new set of GoogleEvent.

After this commit, the addition operator is added for acheving
that, in the most optimized way, which is combining the underlying
sets of events and return a new instance.

task-3809350

Part-of: odoo/odoo#172780
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>