[FIX] base: merge pdfs error expected self
When merging pdfs, if there is an error when meging those pdfs (due to
a malformed PDF for example), the UserError that should be shown to
the user is not due to an error in the arguments given to the
handle_error function.

The aim here is to keep the same function signature for both the function
passed as an argument and _handle_merge_pdfs_error as well as the
possibility to override _handle_merge_pdfs_error.

Discovered during : task-3603619

closes odoo/odoo#211611

Signed-off-by: Krzysztof Magusiak (krma) <krma@odoo.com>