[FIX] mail: show display_name for contact without name and email
Steps to reproduce
===============
1. Create a contact of type invoice address without name and email.
2. Go to any app with chatter.
3. Add this user to the recipient

----> Tag with label `Unnamed` will be shown.

Along with this, if there is any follower without a name and email,
`false` was shown in follower list and in the dialog to edit the
subscription.

After this commit, we will use the display_name as a fallback to show in
those places.

Note: Added/adapted tests to ensure the correct behaviour. Also some
mocked models were missing the definition for computing correct display
name. Thus, compute methods are overriden in those mocked models.

Task-4812554

closes odoo/odoo#253981

X-original-commit: 72aec9c9e348d587a5f413ed607561a6e0f91768
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>