[IMP] marketing_card: set default values on campaigns
Currently if you create a new card campaign the layout is completely
empty until you start filling in fields.

This looks really odd and doesn't give a good idea of what the template
actually looks like. Instead we want there to be some pre-filled values
so the layout is clear to the user.

task-4247003

Part-of: odoo/odoo#201197
Related: odoo/upgrade#7365
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>