[IMP] website_blog: change position of fields in `New Blog Post` wizard
When a user opens the `New Blog Post` dialog, the cursor is initially
focused on the `Select Blog` dropdown field by default, which
interrupts the user flow.
This commit reorders the form fields so that the `Title` field appears
above the `Select Blog` option, ensuring the cursor naturally focuses
on the title input when creating a new blog.

task-4680670

Part-of: odoo/odoo#220029
Signed-off-by: Francois Georis (fge) <fge@odoo.com>