[FIX] point_of_sale: prevent rounding error with combos
**Steps to reproduce:**
- Create a combo, set the free and max to 3
- Create the product combo and set it's price to 50
- Go to PoS, order said combo and chose the same product 3 times
- The price is 49.98 instead of 50

**Problem:**
When ordering a combo and taking the same product multiple times,
the method to adjust the price does not work as it should. In this
case, the method would subtract too much from the total, leading it
to be only 49.98 instead of 50.

**Why the fix:**
This happens because when ordering a combo, each line will get
it's price_unit from the combo's price divided by the number of
lines.
In the case where we have 3 different products, hence 3
different lines, the first two would have a unit_price of 16.67
and the last line will be adjusted as to make it equal to the
combo's price, so the last line would be 16.66, making the sum 50.
But in the case where there is only one line, the last line is
still adjusted, making all the lines 16.66, introducing this error.
The solution would be to take the line's qty into account when computing
how much we should subtract.
Unfortunalty, doing so will, in some cases, introduce a rounding
error. If the new unit_price does not round up well (like in this
case where it is 16.666668 before rounding and 16.67 after),
the rounding will cause a difference between what's shown
in the frontend and what's written on the lines in the
backend, as the lines are rounded up again before
being saved to the backend.
The introduced solution is to split the lines, and each
have a quantity of 1. With this solution, we can have a
different price_unit for each line, which resolves our
problem, as the first two lines will have 16.67 and the
last one will have 16.66 as price_unit. This is how it
was done up until version 18.0, showing every line
with a quantity of 1 instead of grouping them.
As the grouped lines are now single, some rounding
with taxes might differ from what it was before,
which is why a test was altered.

opw-4931215

closes odoo/odoo#252331

X-original-commit: b3af51727073b8cd12391e3a757f1184ef4d9874
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Arthur Nanson (artn) <artn@odoo.com>