[IMP] orm: cache access and record rules
Currently, for each user, we check the database for access and put the
result in the default cache. Since that cache is quite often
invalidated, we redo this for each user of the database. Access rules
change only rarely.

Goal: perform one query for model access and one query for record rules,
and store them in the stable cache. We still use the default cache for
per-user access, but now it can be computed quickly from the stable
cache without additional database queries.

For model access, `_get_all_access_groups` reads the whole table. This
function is then used to implement existing functions that no longer
need to query the database. For a given mode (operation), then model, it
returns the list of group IDs that have access.

For record rules, we cache all rules similarly and store them in a dict
where, for each mode, we have a list of applicable rules. That list is
already pre-parsed for non-dynamic domains to avoid overhead for each
user.

Part-of: odoo/odoo#221598
Signed-off-by: Raphael Collet <rco@odoo.com>