[FIX] l10n_it_stock_ddt,stock_delivery: kit component value
Steps to reproduce:
- Have an IT company setup
- Create a product with a Sales Price and define a kit BOM with 2
  components
- Create SO with product
- Confirm, go to delivery, validate
- Print

Issue:
In the delivery DDT, there is a summary of the delivery where each item
has its own entry (product, quantity, value).
However, in case of kit BOM, each component is reported with the full value
of the sale operation.

Analysis:
This occurs because in the report code we don't consider the possibility
of kit products, where multiple components are associated to the same
sale line.

opw-5013606

closes odoo/odoo#257445

X-original-commit: e4f1d429f4a6cffaa8834959467ea0c3a45d73ee
Signed-off-by: Yolann Sabaux (yosa) <yosa@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>