[IMP] marketing_card{_event}: redirect to record url
If no target url is set and the target record is published we can
redirect to its page to give a more relevant page than simply the home
page of the website.

This would probably be the main way to use the app.

task-4247003

Part-of: odoo/odoo#201197
Related: odoo/upgrade#7365
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>