[IMP] website: add new checkbox specific form options
This commit introduces layout flexibility to easily add checkboxes such
as "Accept terms" or "Send me emails" in forms.

A field will also be "required" by default when the "Checkbox" type is
used, and a marker (*) will be added to the checkbox description when
"required" and its label is hidden.

task-3675302

Part-of: odoo/odoo#223947
Signed-off-by: Outagant Mehdi (mou) <mou@odoo.com>