[IMP] html_builder, *: select field description content when enabled
*: website

When users enable the description option, it’s not immediately clear
how to add content. This causes a short moment of confusion while they
figure out they need to navigate to the form.

- Automatically focus the description input when enabled.

- Select existing content on focus to highlight where to type (thanks
to the floating editor toolbar, it will help to understand the input is
on the left).

task-3675302

Part-of: odoo/odoo#223947
Signed-off-by: Outagant Mehdi (mou) <mou@odoo.com>