[FIX] website_sale: prevent data loss when creating new ribbon
Before this commit, creating a new ribbon using the + icon and changing
options would reset colors and ribbon title to None.

The root cause was that `_setRibbon` returns a promise, but we weren't
waiting for it to complete. This meant that call to `setRibbon` was made
before the actual database ID was assigned to the element, causing
ribbon information to be stored with an incorrect ID.

This commit ensures we await `_setRibbon` before calling `setRibbon`
with the proper database ID. Additionally, the mapping between local
ribbon IDs and server IDs (`localIdToServerId`) has been improved and is
now stored in the component instance for reliable access to the correct
server ID.

task-5503716

closes odoo/odoo#254567

X-original-commit: 01f5db95f95e2e6b9e0d3b4f1d98147cd3124f93
Signed-off-by: Lina Eweis (liew) <liew@odoo.com>