[IMP] loyalty: prevent setting expiration date in the past
Before this commit, it was possible to create or update loyalty cards
with an expiration date in the past, potentially leading to invalid or
unusable coupons.

After this commit, a ValidationError is raised when attempting to set
expiration_date to a past date, ensuring data integrity for loyalty cards.

task-5075177

Part-of: odoo/odoo#228358
Related: odoo/enterprise#102407
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>