[IMP] html_builder, *: use resize observer to detect mobile view switch
*: website

For elements that are invisible based on the size of the window, the
"Invisible Elements" panel did not update its state when the size of the
window changed (but instead it only did when the "Mobile Preview" button
was toggled).

This commit observes the size of the window to trigger the update when
the mobile breakpoint size is crossed.

All the implementations of `on_mobile_preview_clicked_handlers` need to
know if the builder is currently in mobile view, so this commit updates
them all to use the new resource `on_mobile_view_switched_handlers`
which is triggered by the observer of the window size (and remove the
old resource as it is now unused).

Steps to reproduce:
- Open website builder
- Drop a section snippet and click on it
- Set the "Visibility" option to make it invisible on mobile
- Toggle the "Mobile Preview" button
- Observe the snippet becomes hidden and is listed in the
  "Invisible Elements" panel
- Toggle the "Mobile Preview" button back to desktop view
- Reduce the size of the window
- Bug: the snippet becomes hidden, but is not listed in the
  "Invisible Elements" panel, and it is still the target of the builder

task-5149984

closes odoo/odoo#247835

Signed-off-by: Soukéina Bojabza (sobo) <sobo@odoo.com>