[FIX] orm: make with_company() consistent with its documentation
Before this commit, the behavior of method `with_company` was not fully
compliant with its documentation, and could be confusing: it was
sometimes giving more access, sometimes restricting access.

The behavior was the following:
* when `allowed_company_ids` was not set, `with_company` was restricting
  the accepted companies to its argument only.
* when `allowed_company_ids` was set, `with_company` was adding a new
  value to the accepted companies.

For instance:
```python
In [1]: self.env.companies
Out[1]: res.company(3, 4, 1)

In [2]: self.env.context.get('allowed_company_ids')

In [3]: self.with_company(4).env.companies
Out[3]: res.company(4,)

In [4]: self.with_context(allowed_company_ids=[3, 4]).with_company(4).env.companies
Out[4]: res.company(4, 3)
```

This makes the code hardly predictable, because to fully understand the
result of the function, we need to be aware of the context in which the
function is called.

This commit is unifying the behavior to always add a new value to the
allowed companies, and doesn't change the reordering purpose so that the
main company is the first element of `allowed_company_ids`.  In other
words, we always have the following equality (including the order in the
recordsets):

    self.with_company(C).env.companies == C | self.env.companies

Part-of: odoo/odoo#185511
Related: odoo/enterprise#72989
Signed-off-by: Raphael Collet <rco@odoo.com>