[IMP] orm: DomainCustom.optimize
Allow optimizing custom domains. There are some cases where we may want
to have a custom domain for filtering, but for SQL, we could sometimes
translate it back to a normal domain.

The use case: we can use ORM methods to check access, but we can
translate it into the record rules domain.

Part-of: odoo/odoo#254381
Related: odoo/enterprise#110965
Signed-off-by: Raphael Collet <rco@odoo.com>