[IMP] core, *: fields.Binary return BinaryValue
A new class `BinaryValue` represents lazily loaded binary objects. It is
a subclass of `collections.abc.Buffer`. We use it as the cache and
record format of Binary fields. It allows to handle them while delaying
loading of the contents in the memory. At the same time, the loaded
contents are in the raw format which is 33% smaller than base64-encoded
binary values we were storing in cache and that encoding does not have
to be done anymore on the ORM side. We encode in base64 only for RPC
communications.

The `BinaryValue` class has a few helper methods such as getting the
contents (bytes), encoding value to base64, getting mimetype or some
methods like `bytes` object (ie. `decode()`).
To build a value, use `BinaryBytes` (accepts any buffer).

Consequences of changing the return type of Binary fields:
- They all behave the same: we get raw bytes wrapped in a lazy object.
- You can assign directly bytes to the `raw` field, but for other binary
  fields, you have to create a `BinaryValue`. This is because we have no
  way of knowing whether the value was base64-encoded or not.
- base64 encoding should only be used for transfering data in RPC calls.
- Assiging a `str` to a binary field assumes that it is base64-encoded
  and decodes it.
- `bin_size` context key is only used for `Model.read()`. However,
  non-attachment binary fields are fetched entirely now.
- `datas` field is softly deprecated, use `raw` directly.  In future
  work, `datas` field will be removed.
- If you are unsure of the assigned data type, you may cast it to bytes.
- Working with images is a little bit easier.

task-4251301

closes odoo/odoo#244421

Related: odoo/enterprise#104714
Related: odoo/documentation#16093
Signed-off-by: Krzysztof Magusiak (krma) <krma@odoo.com>