[IMP] sale_purchase: make subcontract service option easier to discover
In Odoo, it was already possible to configure a service product to generate a
subcontracting RFQ upon Sales Order confirmation. However, this option
was hard to discover, as it was located in the Purchase tab.

To improve usability, this option has been moved to the 'Create on Order
selection field' in the General Information tab, alongside other similar actions
making it more visible and accessible.

Overall, this enhancement centralises subcontract service configuration for
service-type products and improves clarity and usability.

Related test cases have been updated to reflect the new logic.

Task ID: 5249083

closes odoo/odoo#235670

Related: odoo/enterprise#99481
Related: odoo/upgrade#8850
Signed-off-by: William Henrotin (whe) <whe@odoo.com>