[IMP] project, sale_project: show Customer Rating button only when appropriate
Before this commit:
The Customer Rating button in the project kanban view was only shown based when
sale_timesheet is installed.

When sale_project was installed button appear for non-billable projects.

After this commit:
The Customer Rating button is shown when the project module is installed and
the rating feature is enabled.

If sale_project is installed, the button is shown only when the project is
billable (i.e. rating feature enabled AND project is billable).

task-5364377

closes odoo/odoo#238816

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>