[FIX] html_editor: disable computation on manual btn-outline color
Previously, we created a contrast-adjusted color for the `btn-outline`
classes to ensure readability. This was also applied to manual colors
selected via Theme tab, which made some inconsistencies with standard
`btn`.

This commit fixes that by providing the contrast-adjusted color only on
default palettes, and use the manual color as is when selected in the
Theme tab.

task-5392258

closes odoo/odoo#255479

X-original-commit: c47848be4e239a17cdefaa7cf939d829860324b1
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>
Signed-off-by: Julien de Hosté (jdeh) <jdeh@odoo.com>