[IMP] hr_calendar: Fix double greying of days in calendar.
The calendar app/view uses two different methods to grey out days: the getUnusualDays
method and the businessHours api option from the Fullcalendar library. Ideally,
we would use only one method, but the businessHours API only works reliably in
the "day" and "week" view. It technically greys out days in the "month" and
"year" view, it fails to handle unusual schedules correctly f.ex working only
the first week of the month. That is one of the reason why we also implement
getUnusualDays to grey out specific portions of the month or the year.

However, In the "day" and "week" view both methods apply simultaneously so if
an employee works f.ex FRI-SAT-SUN, all three days are colored white by
the businessHours API but Saturday and Sunday are then colored grey again (in
a slightly different shade) by getUnusualHours (based on another calendar).

This PR aims to rectify that by disabling the unusualDays in "day" and "week"
view, where the businessHours work, and extending unusualDays to also include
the working schedule (if applicable) in "month" and "year" view.

Additionally, the calendar previously failed to properly handle working slots
from a calendar in a radically different timezone from the viewer. F.ex a
working slot that starts "yesterday" from the viewer's perspective would be
ignored. This PR fixes that by detecting if and when a working slots cross
day boundaries and splits them accordingly.

task-5085284

Part-of: odoo/odoo#239420
Related: odoo/enterprise#101790
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>