[CLN] hr_calendar: remove dead code.
In version 18.3, the "Everybody's Calendar" option was removed from the
calendar app. However, much of the code related to this feature was left in
the hr_calendar module. This PR removes all unused logic related to this
non-existent feature.
Additionally, there were a lot of tests and related code that only ran if
a non-existent module was installed so those got either removed or merged
into other test classes, depending on if they were relevant.

task-5085284

closes odoo/odoo#239420

Related: odoo/enterprise#101790
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>