[IMP] purchase: recompute date_planned if overdue
This commit facilitates the re-computation of `date_planned`
for users that didn't confirm the RFQ before the order deadline,
in that case the `date_planned` is no longer valid and needs
to be recomputed. The user is left with the choice to recompute it
or not by clicking the refresh icon next to the `date_planned`

The `date_planned` will be highlighted in red when the date is
less than today's date to warn the user.

Tool tip has been added to the icon to describe what its used for.

Additonally, if there's a product with no lead time, the
information set on the RFQ (dates) will be used for the recompute.

closes odoo/odoo#240678

Task: 5379743
Signed-off-by: William Henrotin (whe) <whe@odoo.com>