[PERF] point_of_sale: fix memory while opening a session.
Before this commit, while opening a session in the POS, all the `product.uom` records were retrieved regardeless of what products were loaded. This might cause a memory error in cases where a lot of records are being retrieved. A solution for this is to retreive the records **only related to the `product.product` records being retrieved**.

opw-5408586

closes odoo/odoo#254673

X-original-commit: 2831861cbff5c5d6326b587ef27e3759c463fb9e
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Mohanned Ahmed (admd) <admd@odoo.com>