[IMP] web, *: improve TextInput on touch and desktop
The commit aims to improve the UX on touch screens, by improving
inputs and buttons design to make them more usable on those touch
oriented devices first by giving more space around certain elements
and making other UI elements more compact to give the user more
information. As a side effect, this task also allowed to improve the
desktop experience regarding suffixes and to unify the look and
feel of text inputs.

A dedicated o_input_box class has been added. Being used at the highest
level on an input/button element or its parent in the UI, it can be used
to add overlay prefix and/or suffix inside of them, using on those UI
elements the 'o_input_box_overlay_ + start/end' classnames.

For example:
<div class='o_input_box'>
    <i class='fa fa-gear o_input_box_overlay_start'/>
    <input class='o_input'/>
    <button class='btn btn-link o_input_box_overlay_end'>Click Me</button>
    <i class='fa fa-arrow o_input_box_overlay_end'/>
</div>

In this example, an input box would be display around the input, with
enough padding before and after the input value, displaying 2 icons
and a button.

Currently, positioning the overlay elements require to call the dedicated
utility function, exported from '/web/static/src/core/input_box.js'.
Padding is then calculated and adapted accordingly to maintain affordance
of the input or button text, and place the dropdown arrow.

An InputBox component has been added as well, implementing easily the right
elements around an input, but it is possible to use the classnames without
the need for this specific component.

If an element with o_input_box is contained inside another o_input_box
parent, the highest element is considered as the input box, and displayed
correctly. This allows a component that already have the class to have a
sibling element serving as an overlay, by adding a parent to both elements
having the o_input_class. It is also useful when some UI is declared from
the arch XML while other overlays are defined from the component side.

This commit also removes the need of a dedicated component for the boolean
toggle field in list views. The edition of any boolean field in
readonly/non selected rows in list views is already well handled in a
scss rule, and it is not necessary to have this dedicated wrapper around
the standard BooleanToggleField component.

The commit also contains other noticable changes and improvements across
Odoo to improve the mobile experience, such as the use of BottomSheet for
example where it could be used instead of a tooltip, or improving the display
of some views using a layout with fields and an image placed on top of the
form, by using a common 'form header' implementation.

Tests might have been dissociated between mobile and desktop, and new tests
have been added as well, to ease the maintenance of the features being
added with this task.

task-5355007

closes odoo/odoo#250051

Related: odoo/enterprise#108283
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Co-authored-by: res <res@odoo.com>