[FIX] product_expiry: removal date is computed after modification
version : saas-18.4+e

Steps to reproduce
------------------
Create a product tracked by lots and enable the expiration date use (use_expiration_date = True).

Create a receipt with this product and mark it as to do.

Open the stock.move.line list view using the "Details" button on the receipt form and try to modify expiration date, it should modify the removal date due to the compute method.

Then, try to modify the removal date and save the modifications. If you reopen the list view, the removal date has been computed again and its value changed.

The fix
------------------
We explicitly set the readonly field to False for the removal_date field from the stock.move.line model in product_expiry.

opw-5368007

closes odoo/odoo#251653

X-original-commit: 08d5ad51d54f09738cac6e02e9291b1cad6ad616
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Samuel Wertz (sawer) <sawer@odoo.com>